revert NDArrayMixin
