+++ xarray/core/indexing.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 import enum
 import functools
-import math
 import operator
 from collections import Counter, defaultdict
 from collections.abc import Hashable, Iterable, Mapping
@@ -27,6 +26,7 @@ from xarray.core.pycompat import (
 )
 from xarray.core.types import T_Xarray
 from xarray.core.utils import (
+    NDArrayMixin,
     either_dict_or_kwargs,
     get_valid_numpy_dtype,
     is_scalar,
@@ -458,52 +458,8 @@ class ExplicitlyIndexed:
         return self.array
 
 
-class ExplicitlyIndexedNDArrayMixin(ExplicitlyIndexed):
+class ExplicitlyIndexedNDArrayMixin(NDArrayMixin, ExplicitlyIndexed):
     __slots__ = ()
-
-    @property
-    def ndim(self: Any) -> int:
-        """
-        Number of array dimensions.
-
-        See Also
-        --------
-        numpy.ndarray.ndim
-        """
-        return len(self.shape)
-
-    @property
-    def size(self: Any) -> int:
-        """
-        Number of elements in the array.
-
-        Equal to ``np.prod(a.shape)``, i.e., the product of the array’s dimensions.
-
-        See Also
-        --------
-        numpy.ndarray.size
-        """
-        return math.prod(self.shape)
-
-    def __len__(self: Any) -> int:
-        try:
-            return self.shape[0]
-        except IndexError:
-            raise TypeError("len() of unsized object")
-
-    @property
-    def dtype(self: Any) -> np.dtype:
-        return self.array.dtype
-
-    @property
-    def shape(self: Any) -> tuple[int, ...]:
-        return self.array.shape
-
-    def __getitem__(self: Any, key):
-        return self.array[key]
-
-    def __repr__(self: Any) -> str:
-        return f"{type(self).__name__}(array={self.array!r})"
 
     def get_duck_array(self):
         key = BasicIndexer((slice(None),) * self.ndim)
@@ -515,7 +471,7 @@ class ExplicitlyIndexedNDArrayMixin(ExplicitlyIndexed):
         return np.asarray(self.get_duck_array(), dtype=dtype)
 
 
-class ImplicitToExplicitIndexingAdapter:
+class ImplicitToExplicitIndexingAdapter(NDArrayMixin):
     """Wrap an array, converting tuples into the indicated explicit indexer."""
 
     __slots__ = ("array", "indexer_cls")
@@ -526,47 +482,6 @@ class ImplicitToExplicitIndexingAdapter:
 
     def __array__(self, dtype: np.typing.DTypeLike = None) -> np.ndarray:
         return np.asarray(self.get_duck_array(), dtype=dtype)
-
-    @property
-    def ndim(self: Any) -> int:
-        """
-        Number of array dimensions.
-
-        See Also
-        --------
-        numpy.ndarray.ndim
-        """
-        return len(self.shape)
-
-    @property
-    def size(self: Any) -> int:
-        """
-        Number of elements in the array.
-
-        Equal to ``np.prod(a.shape)``, i.e., the product of the array’s dimensions.
-
-        See Also
-        --------
-        numpy.ndarray.size
-        """
-        return math.prod(self.shape)
-
-    def __len__(self: Any) -> int:
-        try:
-            return self.shape[0]
-        except IndexError:
-            raise TypeError("len() of unsized object")
-
-    @property
-    def dtype(self: Any) -> np.dtype:
-        return self.array.dtype
-
-    @property
-    def shape(self: Any) -> tuple[int, ...]:
-        return self.array.shape
-
-    def __repr__(self: Any) -> str:
-        return f"{type(self).__name__}(array={self.array!r})"
 
     def get_duck_array(self):
         return self.array.get_duck_array()
