+++ xarray/core/utils.py
@@ -41,6 +41,7 @@ import importlib
 import inspect
 import io
 import itertools
+import math
 import os
 import re
 import sys
@@ -538,6 +539,69 @@ class OrderedSet(MutableSet[T]):
 
     def __repr__(self) -> str:
         return f"{type(self).__name__}({list(self)!r})"
+
+
+class NdimSizeLenMixin:
+    """Mixin class that extends a class that defines a ``shape`` property to
+    one that also defines ``ndim``, ``size`` and ``__len__``.
+    """
+
+    __slots__ = ()
+
+    @property
+    def ndim(self: Any) -> int:
+        """
+        Number of array dimensions.
+
+        See Also
+        --------
+        numpy.ndarray.ndim
+        """
+        return len(self.shape)
+
+    @property
+    def size(self: Any) -> int:
+        """
+        Number of elements in the array.
+
+        Equal to ``np.prod(a.shape)``, i.e., the product of the array’s dimensions.
+
+        See Also
+        --------
+        numpy.ndarray.size
+        """
+        return math.prod(self.shape)
+
+    def __len__(self: Any) -> int:
+        try:
+            return self.shape[0]
+        except IndexError:
+            raise TypeError("len() of unsized object")
+
+
+class NDArrayMixin(NdimSizeLenMixin):
+    """Mixin class for making wrappers of N-dimensional arrays that conform to
+    the ndarray interface required for the data argument to Variable objects.
+
+    A subclass should set the `array` property and override one or more of
+    `dtype`, `shape` and `__getitem__`.
+    """
+
+    __slots__ = ()
+
+    @property
+    def dtype(self: Any) -> np.dtype:
+        return self.array.dtype
+
+    @property
+    def shape(self: Any) -> tuple[int, ...]:
+        return self.array.shape
+
+    def __getitem__(self: Any, key):
+        return self.array[key]
+
+    def __repr__(self: Any) -> str:
+        return f"{type(self).__name__}(array={self.array!r})"
 
 
 class ReprObject:
